LPS-140640 Move files to CP's components dir
