+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/components/BadgePillFilter/index.js
@@ -8,7 +8,7 @@
  * permissions and limitations under the License, including but not limited to
  * distribution rights of the Software.
  */
-import {Button} from '../../../../../../../common/components';
+import {Button} from '../../../../common/components';
 
 const BadgePill = ({filterName, filterValue, onClick}) => {
 	return (
