+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/components/CheckboxFilter/utils/getAvailableFieldsCheckboxs.js
