+++ modules/dxp/apps/osb/osb-site-initializer/osb-site-initializer-customer-portal/extra/remote-app/src/routes/customer-portal/containers/DXPActivationKeysTable/components/BadgeFilter/index.js
@@ -12,8 +12,8 @@
 import {useCallback} from 'react';
 import Button from '../../../../../../common/components/Button';
 import getCurrentEndDate from '../../../../../../common/utils/getCurrentEndDate';
+import BadgePill from '../../../../components/BadgePillFilter';
 import {INITIAL_FILTER} from '../../utils/constants/initialFilter';
-import BadgePill from './components';
 
 const BadgeFilter = ({
 	activationKeysLength,
